api_yamdb/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==2.4.0
djangorestframework_simplejwt==4.8.0
python-dotenv==0.17.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
